requirements.txt
blinker==1.5
click==8.0.3
Flask==2.3.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
Pillow==9.5.0
PyPDF2==3.0.1
pytesseract==0.3.10
Werkzeug==2.3.4
